requirements.txt
PyQt5
PyYAML
bottleneck
gymnasium
hydra-core
inputs
ipykernel
ipython
ipywidgets
matplotlib
matplotlib
numexpr
numpy
pandas
scipy
shapely
stable-baselines3[extra]
tensorflow
tensorrt